refactor(mitm-addon): extract flow metadata key linter
