+++ crates/runner/mitm-addon/tests/fixtures/flow_metadata_key_linter/dynamic_star_args.base.py.txt
@@ -1,0 +1,11 @@
+
+flow.metadata.get(*args, "vm_run_id")
+flow.metadata.update(*args, {"firewall_name": "github"})
+flow.metadata.update(dict.fromkeys(*args, ["firewall_action"], "ALLOW"))
+flow.metadata.get(*(args if condition else other_args), "auth_cache_hit")
+flow.metadata.get(*(args if condition else ["firewall_error"]))
+flow.metadata.update(*([[("firewall_permission", "read")]] if condition else args))
+flow.metadata.get(*[*([prefix] if condition else args), "vm_network_log_path"])
+flow.metadata.get(*args, metadata_keys.VM_RUN_ID)
+flow.metadata.update(*args, {metadata_keys.FIREWALL_NAME: "github"})
+flow.metadata.get(*(args if condition else [metadata_keys.AUTH_CACHE_HIT]))
