+++ crates/runner/mitm-addon/tests/fixtures/flow_metadata_key_linter/violations.py311.py.txt
@@ -1,0 +1,7 @@
+
+except_star_meta = flow.metadata
+try:
+    pass
+except* Exception as except_star_meta:
+    pass
+except_star_meta["firewall_error"] = "auth_failed"
