+++ crates/runner/mitm-addon/tests/fixtures/flow_metadata_key_linter/violations.py312.py.txt
@@ -1,0 +1,30 @@
+
+def function_type_param_meta[T: flow.metadata["vm_run_id"]]():
+    pass
+def function_type_param_merge_meta[T: flow.metadata | {"vm_run_id": "run-1"}]():
+    pass
+async def async_function_type_param_meta[T: flow.metadata["firewall_action"]]():
+    pass
+class ClassTypeParamMeta[T: flow.metadata["firewall_name"]]:
+    pass
+class ClassTypeParamMergeMeta[T: flow.metadata | {"firewall_name": "github"}]:
+    pass
+class ClassKeywordMergeMeta(Base, option=flow.metadata | {"vm_run_id": "run-1"}):
+    pass
+type TypeAliasParamMeta[T: flow.metadata["firewall_base"]] = int
+type TypeAliasParamMergeMeta[T: flow.metadata | {"firewall_base": "https://api.example.com"}] = int
+type TypeAliasValueMeta = flow.metadata["vm_run_id"]
+type TypeAliasMergeMeta = flow.metadata | {"firewall_action": "ALLOW"}
+default_type_param_shadow_meta = flow.metadata
+def default_type_param_reads_outer[default_type_param_shadow_meta](
+    arg=default_type_param_shadow_meta["vm_run_id"],
+):
+    pass
+type_param_restored_meta = flow.metadata
+def type_param_scope_restores_alias[type_param_restored_meta]():
+    pass
+type_param_restored_meta["firewall_name"] = "github"
+function_type_param_global_meta = flow.metadata
+def function_type_param_global_reads_module[function_type_param_global_meta]():
+    global function_type_param_global_meta
+    function_type_param_global_meta["firewall_action"] = "ALLOW"
